Refactored the chatbot API
